onnx_requirements.txt
torch
torchvision
timm
onnx
onnxruntime
numpy
pillow
